+++ angular-workspace/projects/ni/nimble-angular/mapping/user/nimble-mapping-user.directive.ts
@@ -1,4 +1,5 @@
 import { Directive, ElementRef, Input, Renderer2 } from '@angular/core';
+import { NimbleMappingDirective } from '@ni/nimble-angular/mapping/base';
 import { type MappingUser, mappingUserTag } from '@ni/nimble-components/dist/esm/mapping/user';
 import type { MappingUserKey } from '@ni/nimble-components/dist/esm/mapping/base/types';
 
@@ -11,24 +12,16 @@ export { mappingUserTag };
 @Directive({
     selector: 'nimble-mapping-user'
 })
-export class NimbleMappingUserDirective {
-    public get key(): MappingUserKey | undefined {
-        return this.elementRef.nativeElement.key;
-    }
-
-    @Input() public set key(value: MappingUserKey | undefined) {
-        this.renderer.setProperty(this.elementRef.nativeElement, 'key', value);
-    }
-
+export class NimbleMappingUserDirective extends NimbleMappingDirective<MappingUserKey> {
     public get displayName(): string | undefined {
         return this.elementRef.nativeElement.displayName;
     }
 
-    // Renaming because property should have camel casing, but attribute should not
-    // eslint-disable-next-line @angular-eslint/no-input-rename
     @Input('display-name') public set displayName(value: string | undefined) {
         this.renderer.setProperty(this.elementRef.nativeElement, 'displayName', value);
     }
 
-    public constructor(protected readonly renderer: Renderer2, protected readonly elementRef: ElementRef<MappingUser>) {}
+    public constructor(protected readonly renderer: Renderer2, protected readonly elementRef: ElementRef<MappingUser>) {
+        super(renderer, elementRef);
+    }
 }