+++ angular-workspace/projects/ni/nimble-angular/unit/volt/tests/nimble-unit-volt.directive.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+import { NimbleUnitVoltModule } from '../nimble-unit-volt.module';
+
+describe('Nimble volt unit', () => {
+    describe('module', () => {
+        beforeEach(() => {
+            TestBed.configureTestingModule({
+                imports: [NimbleUnitVoltModule]
+            });
+        });
+
+        it('custom element is defined', () => {
+            expect(customElements.get('nimble-unit-volt')).not.toBeUndefined();
+        });
+    });
+});