+++ packages/nimble-components/src/rich-text-mention/users/view/styles.ts
@@ -2,11 +2,8 @@ import { css } from '@microsoft/fast-element';
 import { display } from '@microsoft/fast-foundation';
 import {
     mentionFont,
-    mentionDisabledFontColor,
-    bodyFontColor,
-    bodyFont,
-    bodyDisabledFontColor,
-    mentionFontColor
+    mentionFontColor,
+    mentionDisabledFontColor
 } from '../../../theme-provider/design-tokens';
 
 export const styles = css`
@@ -14,34 +11,14 @@ export const styles = css`
 
     :host {
         box-sizing: border-box;
-        font: ${bodyFont};
-        color: ${bodyFontColor};
-        white-space: pre-wrap;
+        font: ${mentionFont};
     }
 
     .control {
-        font: ${mentionFont};
         color: ${mentionFontColor};
-        display: none;
-    }
-
-    :host([disable-editing]) .control {
-        display: inline;
     }
 
     :host([disabled]) .control {
         color: ${mentionDisabledFontColor};
-    }
-
-    :host([disabled]) {
-        color: ${bodyDisabledFontColor};
-    }
-
-    :host([disable-editing]) {
-        font: ${mentionFont};
-    }
-
-    :host([disable-editing]) slot {
-        display: none;
     }
 `;