+++ packages/nimble-components/src/rich-text/editor/tests/rich-text-editor-mention.spec.ts
@@ -1,7 +1,7 @@
 import { html } from '@microsoft/fast-element';
+import { parameterizeSpec } from '@ni/jasmine-parameterized';
 import { richTextEditorTag, RichTextEditor } from '..';
 import { type Fixture, fixture } from '../../../utilities/tests/fixture';
-import { parameterizeNamedList } from '../../../utilities/tests/parameterized';
 import { RichTextEditorPageObject } from '../testing/rich-text-editor.pageobject';
 import { richTextMentionUsersTag } from '../../../rich-text-mention/users';
 import { mappingUserTag } from '../../../mapping/user';
@@ -842,7 +842,7 @@ describe('RichTextEditorMention', () => {
             }
         ] as const;
 
-        parameterizeNamedList(validMentionNodes, (spec, name, value) => {
+        parameterizeSpec(validMentionNodes, (spec, name, value) => {
             spec(`${name} renders as plain text in editor`, async () => {
                 await appendUserMentionConfiguration(element, [
                     { key: 'user:1', displayName: value.content }
@@ -989,6 +989,16 @@ describe('RichTextEditor user mention via template', () => {
             expect(pageObject.isMentionListboxOpened()).toBeTrue();
         });
     });
+
+    it('should get `span` and expected class name when @ character is added into the editor', async () => {
+        await pageObject.setEditorTextContent('@mention');
+
+        expect(pageObject.getMarkdownRenderedTagNames()).toEqual(['P', 'SPAN']);
+        expect(pageObject.getEditorFirstChildTextContent()).toBe('@mention');
+        expect(pageObject.getEditorLastChildAttribute('class')).toBe(
+            'nimble-mention-view-edit'
+        );
+    });
 });
 
 describe('RichTextEditorMentionListbox', () => {
@@ -1023,7 +1033,7 @@ describe('RichTextEditorMentionListbox', () => {
         });
 
         describe('various wacky strings should display as it is in the mention popup option', () => {
-            parameterizeNamedList(wackyStrings, (spec, name) => {
+            parameterizeSpec(wackyStrings, (spec, name) => {
                 spec(`for ${name}`, async () => {
                     await appendUserMentionConfiguration(element, [
                         { key: 'user:1', displayName: name }