+++ packages/nimble-components/src/table/models/tests/sort-operations.spec.ts
@@ -1,8 +1,8 @@
 import type { Row as TanStackRow, SortingFn } from '@tanstack/table-core';
+import { parameterizeSpec } from '@ni/jasmine-parameterized';
 import { TableColumnSortOperation } from '../../../table-column/base/types';
 import { getTanStackSortingFunction } from '../sort-operations';
 import type { TableFieldValue } from '../../types';
-import { parameterizeNamedList } from '../../../utilities/tests/parameterized';
 
 describe('Sort Operations', () => {
     function createFakeRow(
@@ -65,7 +65,7 @@ describe('Sort Operations', () => {
             ]
         }
     ] as const;
-    parameterizeNamedList(testCases, (spec, name, value) => {
+    parameterizeSpec(testCases, (spec, name, value) => {
         spec(name, () => {
             for (let i = 0; i < value.expectedSortOrder.length; i++) {
                 for (let j = 0; j < value.expectedSortOrder.length; j++) {